refactor: patchTweetIdの切り出し
