+++ app/client/patchTweetId.ts
@@ -1,0 +1,11 @@
+export const patchTweetId = async (contentId: string, tweetId: string) => {
+  const res = await fetch(`${process.env.NEXT_PUBLIC_BASE_URL}/api/posts/tweet_id/${contentId}`, {
+    method: 'PATCH',
+    headers: {
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ tweetId }),
+  })
+  const data: { id: string; error: string } = await res.json()
+  if (res.status !== 200) throw new Error(data.error)
+}
